- javafx moved to 21 - java moved to 17 - demo moved to test
