+++ src/test/java/org/dockfx/DockNodeTest.java
@@ -1,0 +1,11 @@
+package org.dockfx;
+
+
+import org.junit.jupiter.api.Test;
+
+class DockNodeTest {
+
+    @Test
+    void setFloating() {
+    }
+}
